aded vehicles feature

+++ src/module/users/user.service.ts
@@ -1,5 +1,5 @@
 // import { pool } from "../../config/db";
-import bcrypt from 'bcryptjs'
+import bcrypt from "bcryptjs";
 import { pool } from "../../config/db";
 
 const getUser = async () => {
@@ -10,23 +10,29 @@ const getUser = async () => {
   return result;
 };
 
-
-
 const getSingleUser = async (id: string) => {
   const result = await pool.query(`SELECT * FROM users WHERE id = $1`, [id]);
   return result;
 };
 
-const updateUser = async (payload:Record<string,unknown>, id: string) => {
-  const {name,email,phone,role,password}= payload
+const updateUser = async (payload: Record<string, unknown>, id: string) => {
+  const { name, email, phone, role, password } = payload;
   const result = await pool.query(
     `UPDATE users SET name =$1, email=$2, phone=$3,role=$4, password=$5 WHERE id=$6 RETURNING *`,
-    [name, email,phone,role,password,id],
+    [name, email, phone, role, password, id],
   );
   return result;
 };
 
 const deleteUser = async (id: string) => {
+  const bookingCheck = await pool.query(
+    `SELECT customer_id FROM bookings WHERE customer_id=$1 AND status ='active'`,
+    [id],
+  );
+  if (bookingCheck.rows.length > 0) {
+    throw new Error("This user has active bookings");
+  }
+
   const result = await pool.query(`DELETE FROM users WHERE id =$1`, [id]);
   return result;
 };
